# conference/src/docs/openapi.yml
openapi: 3.0.1
info:
  title: Conference GO!
  description: 'This is the Professional Develpoment monolith application used to refactor into microservices.'
  contact:
    email: profdev@hackreactor.com
  version: 1.0.0
servers:
- url: http://localhost:8080/
...
paths:
...
  /api/events/{eventId}/presentations:
    parameters:
    - name: eventId
      in: path
      required: true
      description: The numerical identifier of the event
      schema:
        type: integer
    get:
      tags:
      - presentations
      summary: Get the presentation proposals in the system
      responses:
        200:
          description: The list of presentation proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Presentation'
        401:
          description: Unauthorized access with code INVALID_TOKEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
    post:
      tags:
      - presentations
      summary: Create a new presentation proposal in the system
      requestBody:
        description: The presentation proposal
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresentationCreation'
      responses:
        201:
          description: The list of presentation proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Presentation'
        401:
          description: Unauthorized access with code INVALID_TOKEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /api/events/{eventId}/presentations/{id}/approved:
    parameters:
    - name: eventId
      in: path
      required: true
      description: The numerical identifier of the event
      schema:
        type: integer
    - name: id
      in: path
      required: true
      description: The numerical identifier of the presentation
      schema:
        type: integer
    put:
      tags:
      - presentations
      summary: Approves the presentation proposal
      responses:
        200:
          description: The presentation detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationDecided'
        401:
          description: Unauthorized access with code INVALID_TOKEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Could not find a presentation that could be rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /api/events/{eventId}/presentations/{id}/rejected:
    parameters:
    - name: eventId
      in: path
      required: true
      description: The numerical identifier of the event
      schema:
        type: integer
    - name: id
      in: path
      required: true
      description: The numerical identifier of the event
      schema:
        type: integer
    put:
      tags:
      - presentations
      summary: Approves the presentation proposal
      responses:
        200:
          description: The presentation detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationDecided'
        401:
          description: Unauthorized access with code INVALID_TOKEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Could not find a presentation that could be rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
...
components:
  schemas:
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
          enum:
          - BAD_CREDENTIALS
          - INVALID_TOKEN
          - INVALID_IDENTIFIER
          - INVALID_PARAMETER
          - MISSING_PARAMETER
          - VERSION_CONFLICT
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: object
            properties:
              message:
                type: string
              rule:
                type: string
...
    PresentationCreation:
      required:
      - email
      - presenterName
      - companyName
      - title
      - synopsis
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 100
        presenterName:
          type: string
          minLength: 1
          maxLength: 100
        companyName:
          type: string
          minLength: 1
          maxLength: 100
        title:
          type: string
          minLength: 8
          maxLength: 100
        synopsis:
          type: string
          minLength: 50
    Presentation:
      allOf:
      - $ref: '#/components/schemas/PresentationCreation'
      - type: object
        required:
        - id
        - created
        - status
        properties:
          id:
            type: integer
          created:
            type: string
            format: date-time
          status:
            type: string
            enum:
            - SUBMITTED
            - APPROVED
            - REJECTED
    PresentationDecided:
      allOf:
        - $ref: '#/components/schemas/Presentation'
        - type: object
          required:
          - decider
          - decidedOn
          properties:
            decider:
              type: object
              required:
              - name
              - decidedOn
              properties:
                name:
                  type: string
                decidedOn:
                  type: string
                  format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 